61. Вывести трехзначные числа с определенным условием

+++ src/main/java/ru/job4j/loop/Task61.java
@@ -6,8 +6,7 @@ public class Task61 {
     public static void loop() {
         StringJoiner sj = new StringJoiner(" ");
         for (int i = 100; i < 500; i++) {
-            if (Math.pow((double) i / 100, 2) + Math.pow(i % 10, 2)
-                    == Math.pow((double) (i % 100) / 10, 2)) {
+            if (Math.pow(i / 100, 2) + Math.pow(i % 10, 2) == Math.pow(i % 100 / 10, 2)) {
                 sj.add(String.valueOf(i));
             }
         }
